backend/requirements.txt
flask==3.0.3
bayoo-docx==0.2.20